Finished Folder Page and configured Navigation
Created various components and a style sheet to set up the grid that will be used for both the Folder Page and the Notes Page. Altered previous components and styling for greater adaptability and flexbility in future components. Set up navigation using react router dom between the Landing Page and the folder page